pyproject.toml
[project]
name = "collector-watcher"
version = "0.1.0"
description = "Automated monitoring system for OpenTelemetry Collector components"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "PyYAML>=6.0.1",
    "GitPython>=3.1.40",
    "PyGithub>=2.1.1",
    "pytest>=7.4.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.13.3",
]
